requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
python-nmap==0.7.1
pyyaml==6.0.1
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
ldap3==2.9.1
ansible==9.0.1
jinja2==3.1.2
werkzeug==3.0.1